+++ resources/compiledJS/arrays/characterInteractions.js
@@ -175,6 +175,26 @@ const characterInteractions = {
                 ]
             }
         ],
+    },
+    blacksmithMaroch: {
+        always: [
+            {
+                name: "blacksmithMaroch_store",
+                type: "openStore",
+                action: { type: "store", id: "blacksmithMaroch_store_normal" }
+            },
+            {
+                name: "blacksmithMaroch_smithing",
+                type: "smithScreen",
+                action: { type: "smith", id: "blacksmithMaroch_smith" }
+            },
+            {
+                name: "leaveDialog",
+                type: "exit",
+                action: { type: "exit" },
+                displayAtBottom: true,
+            }
+        ]
     }
 };
 //# sourceMappingURL=characterInteractions.js.map