+++ script/hotbar.ts
@@ -644,8 +644,8 @@ function effectSyntax(effect: any, embed: boolean = false, effectId: string = ""
   if (!img) img = icons[key_ + tailEnd + "_icon"];
   if (!img) img = icons[key_ + "_icon"];
   if (value < 0) {
-    text += `§${embed ? " " : ""}<c>${flipColor ? "lime" : "red"}<c><f>${embed ? "15px" : "18px"}<f>${lang["decreases"]}  <i>${frontImg === "" ? img : frontImg}<i>${key} ${backImg ? backImg : ""}${tailEnd} ${lang["by"]}${rawKey.endsWith("P") ? value + "%" : value} ${lastBit}\n`;
-  } else text += `§${embed ? " " : ""}<c>${flipColor ? "red" : "lime"}<c><f>${embed ? "15px" : "18px"}<f>${lang["increases"]} <i>${frontImg === "" ? img : frontImg}<i>${key} ${backImg ? backImg : ""}${tailEnd} ${lang["by"]}${rawKey.endsWith("P") ? value + "%" : value} ${lastBit}\n`;
+    text += `§${embed ? " " : ""}<c>${flipColor ? "lime" : "red"}<c><f>${embed ? "15px" : "18px"}<f>${lang["decreases"]}  <i>${frontImg === "" ? img : frontImg}<i>${key} ${backImg ? backImg : ""}${tailEnd} ${lang["by"]}${rawKey.endsWith("P") ? value.toFixed(1) + "%" : value.toFixed(1)} ${lastBit}\n`;
+  } else text += `§${embed ? " " : ""}<c>${flipColor ? "red" : "lime"}<c><f>${embed ? "15px" : "18px"}<f>${lang["increases"]} <i>${frontImg === "" ? img : frontImg}<i>${key} ${backImg ? backImg : ""}${tailEnd} ${lang["by"]}${rawKey.endsWith("P") ? value.toFixed(1) + "%" : value.toFixed(1)} ${lastBit}\n`;
   return text;
 }
 
@@ -897,6 +897,16 @@ function renderCharacter() {
   coreResistances.classList.add("coreResistances");
   statusResistances.classList.add("statusResistances");
   passiveAbilities.classList.add("passiveAbilities");
+  let allMods = "<f>20px<f>All modifiers: §\n";
+  Object.entries(player.allModifiers).map((eff: any) => {
+    if (eff[1] - 1 === 0) return;
+    else if (eff[1] === 0) return;
+    console.log(eff[1]);
+    const displayEff = { ...eff };
+    if (displayEff[0].endsWith("P") && !displayEff[0].includes("crit")) displayEff[1] = displayEff[1] * 100 - 100;
+    allMods += effectSyntax(displayEff);
+  });
+  tooltip(pc, allMods);
   Object.entries(playerCoreStats).forEach((stat: any) => {
     coreStats.append(createStatDisplay(stat));
   });