+++ script/player.ts
@@ -272,6 +272,7 @@ class PlayerCharacter extends Character {
         }
       });
       this[slot] = {};
+      player.allModifiers = getAllModifiersOnce(player, true);
       renderInventory();
     };
 
@@ -314,6 +315,7 @@ class PlayerCharacter extends Character {
         }
         else this.unequip(event, "offhand", item.index, true, fromContextMenu);
       }
+      player.allModifiers = getAllModifiersOnce(player, true);
       if (!auto) renderInventory();
     };
 
@@ -519,6 +521,7 @@ function sortInventory(category: string, reverse: boolean, inventory: Array<any>
   }
   else inventory.sort((a, b) => numberSort(a, b, category, reverse));
   if (context.includes("SELLING")) createMerchantWindow(false, true);
+  else if (context == "UPGRADE") createSmithingWindow(false);
   else renderInventory();
 }