+++ script/startGame.ts
@@ -211,14 +211,17 @@ function beginGame() {
   turnOver = true;
   enemiesHadTurn = 0;
   lootedChests = [];
+  currentMap = 2;
   state.inCombat = false;
   resetAllChests();
   handleEscape();
   createStaticMap();
   modifyCanvas();
+  renderMinimap(maps[currentMap]);
+  moveMinimap();
   setTimeout(() => {
     openLevelingScreen();
-  }, 100);
+  }, 50);
 }
 
 function checkIfCanStartGame() {
@@ -377,7 +380,7 @@ function classTT(data: any) {
   return txt;
 }
 
-for (let i = 0; i < 20; i++) {
+for (let i = 0; i < 80; i++) {
   player.addItem({ ...randomProperty(items) });
 }
 
@@ -402,3 +405,6 @@ function initGame() {
 }
 
 initGame();
+
+
+createSmithingWindow();